refactor: e2e test cases
