+++ packages/tests/src/e2e/feature/accountCommand.tests.ts
