+++ packages/tests/src/e2e/feature/helpInfo.tests.ts
