--- packages/tests/src/e2e/scaffold/OfficeAddinScaffold.tests.ts
@@ -1,53 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-/**
- * @author Huajie Zhang <huajiezhang@microsoft.com>
- */
-
-import { it } from "@microsoft/extra-shot-mocha";
-import { ProgrammingLanguage } from "@microsoft/teamsfx-core";
-import { expect } from "chai";
-import { describe } from "mocha";
-import path from "path";
-import { Cleaner } from "../../commonlib/cleaner";
-import { Capability } from "../../utils/constants";
-import { Executor } from "../../utils/executor";
-import { getTestFolder, getUniqueAppName } from "../commonUtils";
-
-describe("Office Addin TaskPane Scaffold", function () {
-  let testFolder: string;
-  let appName: string;
-  let projectPath: string;
-
-  // Should succeed on the 3rd try
-  this.retries(2);
-
-  beforeEach(() => {
-    testFolder = getTestFolder();
-    appName = getUniqueAppName();
-    projectPath = path.resolve(testFolder, appName);
-  });
-
-  afterEach(async () => {
-    await Cleaner.clean(testFolder);
-  });
-
-  it(
-    `taskpane`,
-    { testPlanCaseId: 17132789, author: "huajiezhang@microsoft.com" },
-    async function () {
-      {
-        //Temporarily comment test cases and refine it after release process is finished
-        // const result = await Executor.createProject(
-        //   testFolder,
-        //   appName,
-        //   Capability.TaskPane,
-        //   ProgrammingLanguage.TS
-        // );
-        // expect(result.success).to.be.true;
-        expect(true).to.be.true;
-      }
-    }
-  );
-});
